Twaked the damage made by the knive
The knive deals now only half the damage. This should encourage the players to melee instead.

+++ Assets/Scripts/NPC/Monster/MonsterControllerAbstract.cs
@@ -33,7 +33,7 @@ public abstract class MonsterControllerAbstract : NPCControllerAbstract
         SpawnEnemies();
     }
 
-    protected override void Damage(int value){
+    protected override void Damage(float value){
         base.Damage(value);
         StartCoroutine(OnHit());
     }
